Add customizable paths for portfolio config file
There was an issue where a windows user could make the portfolio feature work because of the
hardcoded linux directory. Thus we made the portfolio directory customizable with fallback
address the hardcoded linux directory. Maybe we should identify the OS of the user and
set different fallback dirs for different OS. We'll see what the future holds.